move device starting state to kernel
